backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
openai==1.57.4
python-dotenv==1.0.1
requests==2.32.3
